requirement.txt
opencv-python
Pillow
ultralytics
pandas
tifffile
tqdm